create hadoop module
